==> 07_CNVR/combined/01_combine_models.R <==
# Overlap CNVRs from different models (All).

#################################################
### Libraries ###################################
library(data.table)
library(dplyr)


#################################################
### STEP 1: Load data ###########################

#### Mirror #####################################

# All; File generated by "07_CNVR/mirror/01_CNVR_mirror.R"
mirror <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/mirror/data/GWAS_CNVR_All_mirror.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
mirror$CHR <- as.character(mirror$CHR)
mirror$MAIN_MODEL <- "M"

# Males; File generated by "07_CNVR/mirror/01_CNVR_mirror.R"
mirror_M <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/mirror/data/GWAS_CNVR_M_mirror.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
mirror_M$CHR <- as.character(mirror_M$CHR)
mirror_M$MAIN_MODEL <- "M"
mirror_M <- mirror_M[mirror_M$PHENO %in% c("PC"), ]
if(nrow(mirror_M) > 0) {mirror <- rbind(mirror, mirror_M)}; rm(mirror_M)

# Females; File generated by "07_CNVR/mirror/01_CNVR_mirror.R"
mirror_F <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/mirror/data/GWAS_CNVR_F_mirror.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
mirror_F$CHR <- as.character(mirror_F$CHR)
mirror_F$MAIN_MODEL <- "M"
mirror_F <- mirror_F[mirror_F$PHENO %in% c("BC", "OC", "menstruation", "endometriosis"), ]
if(nrow(mirror_F) > 0) {mirror <- rbind(mirror, mirror_F)}; rm(mirror_F)


#### Duplication-only ###########################

# All; File generated by "07_CNVR/duplication_only/01_CNVR_dup.R"
dup <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/duplication_only/data/GWAS_CNVR_All_dup.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
dup$CHR <- as.character(dup$CHR)
dup$MAIN_MODEL <- "DUP"

# Males; File generated by "07_CNVR/duplication_only/01_CNVR_dup.R"
dup_M <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/duplication_only/data/GWAS_CNVR_M_dup.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
dup_M$CHR <- as.character(dup_M$CHR)
dup_M$MAIN_MODEL <- "DUP"
dup_M <- dup_M[dup_M$PHENO %in% c("PC"), ]
if(nrow(dup_M) > 0) {dup <- rbind(dup, dup_M)}; rm(dup_M)

# Females; File generated by "07_CNVR/duplication_only/01_CNVR_dup.R"
dup_F <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/duplication_only/data/GWAS_CNVR_F_dup.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
dup_F$CHR <- as.character(dup_F$CHR)
dup_F$MAIN_MODEL <- "DUP"
dup_F <- dup_F[dup_F$PHENO %in% c("BC", "OC", "menstruation", "endometriosis"), ]
if(nrow(dup_F) > 0) {dup <- rbind(dup, dup_F)}; rm(dup_F)


#### Deletion-only ############################

# Deletion-only; File generated by "07_CNVR/deletion_only/01_CNVR_del.R"
del <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/deletion_only/data/GWAS_CNVR_All_del.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
del$CHR <- as.character(del$CHR)
del$MAIN_MODEL <- "DEL"

# Males; File generated by "07_CNVR/deletion_only/01_CNVR_del.R"
del_M <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/deletion_only/data/GWAS_CNVR_M_del.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
del_M$CHR <- as.character(del_M$CHR)
del_M$MAIN_MODEL <- "DEL"
del_M <- del_M[del_M$PHENO %in% c("PC"), ]
if(nrow(del_M) > 0) {del <- rbind(del, del_M)}; rm(del_M)

# Females; File generated by "07_CNVR/deletion_only/01_CNVR_del.R"
del_F <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/deletion_only/data/GWAS_CNVR_F_del.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
del_F$CHR <- as.character(del_F$CHR)
del_F$MAIN_MODEL <- "DEL"
del_F <- del_F[del_F$PHENO %in% c("BC", "OC", "menstruation", "endometriosis"), ]
if(nrow(del_F) > 0) {del <- rbind(del, del_F)}; rm(del_F)


#### U-shape ##################################

# All; File generated by "07_CNVR/u_shape/01_CNVR_U.R"
u <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/u_shape/data/GWAS_CNVR_All_U.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
u$CHR <- as.character(u$CHR)
u$MAIN_MODEL <- "U"

# Males; File generated by "07_CNVR/u_shape/01_CNVR_U.R"
u_M <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/u_shape/data/GWAS_CNVR_M_U.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
u_M$CHR <- as.character(u_M$CHR)
u_M$MAIN_MODEL <- "U"
u_M <- u_M[u_M$PHENO %in% c("PC"), ]
if(nrow(u_M) > 0) {u <- rbind(u, u_M)}; rm(u_M)

# Females; File generated by "07_CNVR/u_shape/01_CNVR_U.R"
u_F <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/u_shape/data/GWAS_CNVR_F_U.txt", select = c(7, 1:6, 8:9), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR")))
u_F$CHR <- as.character(u_F$CHR)
u_F$MAIN_MODEL <- "U"
u_F <- u_F[u_F$PHENO %in% c("BC", "OC", "menstruation", "endometriosis"), ]
if(nrow(u_F) > 0) {u <- rbind(u, u_F)}; rm(u_F)



#################################################
### STEP 2: U-shape associations ################

# Start with adding all U-shape associations
df <- u
df$ALL_MODELS <- "U"
df$SIGN_MIRROR <- NA
print(paste0("Number of associations U-shape: ", nrow(df)))



#################################################
### STEP 3: Duplication-only associations #######

print(paste0("Number of duplication associations to add: ", nrow(dup)))

# Loop over duplication-only signals
for (i in 1:nrow(dup)) {

	# Define signal
	pheno <- dup[i, "PHENO"]
	chr <- dup[i, "CHR"]
	start <- dup[i, "START_CNVR"]
	end <- dup[i, "END_CNVR"]

	# Identify possible matches (with data.table::foverlaps)
	df_dup <- data.table(PHENO = pheno, CHR = chr, START_CNVR = start, END_CNVR = end)
	setkey(df_dup, PHENO, CHR, START_CNVR, END_CNVR)
	match <- na.omit(foverlaps(data.table(df), df_dup, type = "any", which = T))
	print(paste0("Number of matches for ", pheno, " & chr", chr, ":", start, "-", end, ": ", nrow(match)))

	# If there are multiple matches, add a warrning
	if (nrow(match) > 1) {print("Warning! Multiple matches... All associations were grouped.")  

		# Create a dataframe with the new row
		df_new <- data.frame(PHENO = dup[i, "PHENO"],
						 	 TOP_SNP = dup[i, "TOP_SNP"], CHR = dup[i, "CHR"], POS = dup[i, "POS"],
						 	 OR = dup[i, "OR"], SE = dup[i, "SE"], P = dup[i, "P"],
						 	 START_CNVR = min(c(df[match$xid, "START_CNVR"], dup[i, "START_CNVR"])), 
							 END_CNVR = max(c(df[match$xid, "END_CNVR"], dup[i, "END_CNVR"])),
						 	 MAIN_MODEL = "DUP", 
							 ALL_MODELS = paste0(paste(unique(unlist(strsplit(df[match$xid, "ALL_MODELS"], "-"))), collapse  = "-"), "-DUP"), 
							 SIGN_MIRROR = NA)

		# Add the new row to the main dataframe & remove the old matching rows
		df <- rbind(df, df_new)
		df <- df[-match$xid, ]}

	# If there is a match: Update the row in the main dataframe
	if (nrow(match) == 1) {

		# Identify the row to modify in the main dataframe
		row_to_mod <- as.numeric(match[1,1])

		# If the duplication-only signal is more significant, change the top signal
		if (dup[i, "P"] <= df[row_to_mod, "P"]) {
			df[row_to_mod, "MAIN_MODEL"] <- "DUP" 
			df[row_to_mod, "TOP_SNP"] <- dup[i, "TOP_SNP"] 
			df[row_to_mod, "POS"] <- dup[i, "POS"] 
			df[row_to_mod, "OR"] <- dup[i, "OR"] 
			df[row_to_mod, "SE"] <- dup[i, "SE"] 
			df[row_to_mod, "P"] <- dup[i, "P"]} 

		# Complete genetic coordinates (maximal CNVR)
		df[row_to_mod, "START_CNVR"] <- min(df[row_to_mod, "START_CNVR"], dup[i, "START_CNVR"])
		df[row_to_mod, "END_CNVR"] <- max(df[row_to_mod, "END_CNVR"], dup[i, "END_CNVR"])
		df[row_to_mod, "ALL_MODELS"] <- paste(unique(c(unlist(strsplit(df[row_to_mod, "ALL_MODELS"],  "-")), "DUP")), collapse  = "-")
	}

	# If there is no match: Add a new row to the main dataframe
	if (nrow(match) == 0) {

		# Create a df with the new row
		df_new <- data.frame(PHENO = dup[i, "PHENO"],
						 	 TOP_SNP = dup[i, "TOP_SNP"], CHR = dup[i, "CHR"], POS = dup[i, "POS"],
						 	 OR = dup[i, "OR"], SE = dup[i, "SE"], P = dup[i, "P"],
						 	 START_CNVR = dup[i, "START_CNVR"], END_CNVR = dup[i, "END_CNVR"],
						 	 MAIN_MODEL = "DUP", ALL_MODELS = "DUP", SIGN_MIRROR = NA)

		# Add the new row to the main dataframe
		df <- rbind(df, df_new)
	}
}

print(paste0("Number of associations U-shape + Duplication-only: ", nrow(df)))



#################################################
### STEP 4: Deletion-only associations #######

print(paste0("Number of deletion associations to add: ", nrow(del)))

# Loop over deletion-only signals
for (i in 1:nrow(del)) {

	# Define signal
	pheno <- del[i, "PHENO"]
	chr <- del[i, "CHR"]
	start <- del[i, "START_CNVR"]
	end <- del[i, "END_CNVR"]

	# Identify possible matches (with data.table::foverlaps)
	df_del <- data.table(PHENO = pheno, CHR = chr, START_CNVR = start, END_CNVR = end)
	setkey(df_del, PHENO, CHR, START_CNVR, END_CNVR)
	match <- na.omit(foverlaps(data.table(df), df_del, type = "any", which = T))
	print(paste0("Number of matches for ", pheno, " & chr", chr, ":", start, "-", end, ": ", nrow(match)))

	# If there are multiple matches, add a warrning
	if (nrow(match) > 1) {print("Warning! Multiple matches... All associations were grouped.")  

		# Create a dataframe with the new row
		df_new <- data.frame(PHENO = del[i, "PHENO"],
						 	 TOP_SNP = del[i, "TOP_SNP"], CHR = del[i, "CHR"], POS = del[i, "POS"],
						 	 OR = del[i, "OR"], SE = del[i, "SE"], P = del[i, "P"],
						 	 START_CNVR = min(c(df[match$xid, "START_CNVR"], del[i, "START_CNVR"])), 
							 END_CNVR = max(c(df[match$xid, "END_CNVR"], del[i, "END_CNVR"])),
						 	 MAIN_MODEL = "DEL", 
							 ALL_MODELS = paste0(paste(unique(unlist(strsplit(df[match$xid, "ALL_MODELS"], "-"))), collapse  = "-"), "-DEL"), 
							 SIGN_MIRROR = NA)

		# Add the new row to the main dataframe & remove the old matching rows
		df <- rbind(df, df_new)
		df <- df[-match$xid, ]}

	# If there is a match: Update the row in the main dataframe
	if (nrow(match) == 1) {

		# Identify the row to modify in the main dataframe
		row_to_mod <- as.numeric(match[1,1])

		# If the  deletion-only signal is more significant, change the top signal
		if (del[i, "P"] <= df[row_to_mod, "P"]) {
			df[row_to_mod, "MAIN_MODEL"] <- "DEL" 
			df[row_to_mod, "TOP_SNP"] <- del[i, "TOP_SNP"] 
			df[row_to_mod, "POS"] <- del[i, "POS"] 
			df[row_to_mod, "OR"] <- del[i, "OR"] 
			df[row_to_mod, "SE"] <- del[i, "SE"] 
			df[row_to_mod, "P"] <- del[i, "P"]} 

		# Complete genetic coordinates (maximal CNVR)
		df[row_to_mod, "START_CNVR"] <- min(df[row_to_mod, "START_CNVR"], del[i, "START_CNVR"])
		df[row_to_mod, "END_CNVR"] <- max(df[row_to_mod, "END_CNVR"], del[i, "END_CNVR"])
		df[row_to_mod, "ALL_MODELS"] <- paste(unique(c(unlist(strsplit(df[row_to_mod, "ALL_MODELS"],  "-")), "DEL")), collapse  = "-")
	}

	# If there is no match: Add a new row to the main dataframe
	if (nrow(match) == 0) {

		# Create a df with the new row
		df_new <- data.frame(PHENO = del[i, "PHENO"],
						 	 TOP_SNP = del[i, "TOP_SNP"], CHR = del[i, "CHR"], POS = del[i, "POS"],
						 	 OR = del[i, "OR"], SE = del[i, "SE"], P = del[i, "P"],
						 	 START_CNVR = del[i, "START_CNVR"], END_CNVR = del[i, "END_CNVR"],
						 	 MAIN_MODEL = "DEL", ALL_MODELS = "DEL", SIGN_MIRROR = NA)

		# Add the new row to the main dataframe
		df <- rbind(df, df_new)
	}
}

print(paste0("Number of associations U-shape + Duplication-only + Deletion-only: ", nrow(df)))



#################################################
### STEP 5: Mirror associations #################

print(paste0("Number of mirror associations to add: ", nrow(mirror)))

# Loop over mirror signals
for (i in 1:nrow(mirror)) {

	# Define signal
	pheno <- mirror[i, "PHENO"]
	chr <- mirror[i, "CHR"]
	start <- mirror[i, "START_CNVR"]
	end <- mirror[i, "END_CNVR"]
	if(mirror[i, "OR"] < 1) {sign <- "-"} else {sign <- "+"}

	# Identify possible matches (with data.table::foverlaps)
	df_mirror <- data.table(PHENO = pheno, CHR = chr, START_CNVR = start, END_CNVR = end)
	setkey(df_mirror, PHENO, CHR, START_CNVR, END_CNVR)
	match <- na.omit(foverlaps(data.table(df), df_mirror, type = "any", which = T))
	print(paste0("Number of matches for ", pheno, " & chr", chr, ":", start, "-", end, ": ", nrow(match)))

	# If there are multiple matches, add a warrning
	if (nrow(match) > 1) {print("Warning! Multiple matches... All associations were grouped.")  

		# Create a dataframe with the new row
		df_new <- data.frame(PHENO = mirror[i, "PHENO"],
						 	 TOP_SNP = mirror[i, "TOP_SNP"], CHR = mirror[i, "CHR"], POS = mirror[i, "POS"],
						 	 OR = mirror[i, "OR"], SE = mirror[i, "SE"], P = mirror[i, "P"],
						 	 START_CNVR = min(c(df[match$xid, "START_CNVR"], mirror[i, "START_CNVR"])), 
							 END_CNVR = max(c(df[match$xid, "END_CNVR"], mirror[i, "END_CNVR"])),
						 	 MAIN_MODEL = "M", 
							 ALL_MODELS = paste0(paste(unique(unlist(strsplit(df[match$xid, "ALL_MODELS"], "-"))), collapse  = "-"), "-M"), 
							 SIGN_MIRROR = sign)

		# Add the new row to the main dataframe & remove the old matching rows
		df <- rbind(df, df_new)
		df <- df[-match$xid, ]}

	# If there is a match: Update the row in the main dataframe
	if (nrow(match) == 1) {

		# Identify the row to modify in the main dataframe
		row_to_mod <- as.numeric(match[1,1])

		# If the  mirror signal is more significant, change the top signal
		if (mirror[i, "P"] < df[row_to_mod, "P"]) {
			df[row_to_mod, "MAIN_MODEL"] <- "M" 
			df[row_to_mod, "TOP_SNP"] <- mirror[i, "TOP_SNP"] 
			df[row_to_mod, "POS"] <- mirror[i, "POS"] 
			df[row_to_mod, "OR"] <- mirror[i, "OR"] 
			df[row_to_mod, "SE"] <- mirror[i, "SE"] 
			df[row_to_mod, "P"] <- mirror[i, "P"]} 

		# Complete genetic coordinates (maximal CNVR)
		df[row_to_mod, "START_CNVR"] <- min(df[row_to_mod, "START_CNVR"], mirror[i, "START_CNVR"])
		df[row_to_mod, "END_CNVR"] <- max(df[row_to_mod, "END_CNVR"], mirror[i, "END_CNVR"])
		df[row_to_mod, "ALL_MODELS"] <- paste(unique(c(unlist(strsplit(df[row_to_mod, "ALL_MODELS"],  "-")), "M")), collapse  = "-")
		df[row_to_mod, "SIGN_MIRROR"] <- sign

	}

	# If there is no match: Add a new row to the main dataframe
	if (nrow(match) == 0) {

		# Create a df with the new row
		df_new <- data.frame(PHENO = mirror[i, "PHENO"],
						 	 TOP_SNP = mirror[i, "TOP_SNP"], CHR = mirror[i, "CHR"], POS = mirror[i, "POS"],
						 	 OR = mirror[i, "OR"], SE = mirror[i, "SE"], P = mirror[i, "P"],
						 	 START_CNVR = mirror[i, "START_CNVR"], END_CNVR = mirror[i, "END_CNVR"],
						 	 MAIN_MODEL = "M", ALL_MODELS = "M", SIGN_MIRROR = sign)

		# Add the new row to the main dataframe
		df <- rbind(df, df_new)
	}
}

print(paste0("Number of associations U-shape + Deletion-only + Deletion-only + Mirror: ", nrow(df)))



#################################################
### STEP 6: Corroborating evidence ##############

# Number of association with >2 lines of evidence
print(paste0("Number of associations with evidence from at least two models: ", nrow(df[grep("-", df$ALL_MODELS), ])))

# Table: main model
print("Association  table (main model):")
print(table(df$MAIN_MODEL))

# Table: all models
print("Association  table (all models):")
print(table(df$ALL_MODELS))


 
#################################################
### Save ########################################

fwrite(df, paste0("/cauwerx/gwas/07_CNVR/combined/data/combined_CNVR/all_combined_CNVR.txt"), row.names = F, col.names = F, quote = F, sep = "\t")
