==> 07_CNVR/combined/05_define_statistical_tiers.R <==
# Annotate associations with results of replication/corroborating analyses

#################################################
### Libraries & parameters ######################
library(data.table)
library(dplyr)

# Define the replication threshold
thr <- 1e-4



#################################################
### STEP 1: Load data ###########################

# All merged associations (main logistic regression analysis); File generated by "07_CNVR/combined/01_combine_models.R"
cnvrs <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/combined/data/combined_CNVR/all_combined_CNVR.txt", header = F, select = c(1:11), col.names = c("PHENO", "TOP_SNP", "CHR", "POS", "OR", "SE", "P", "START_CNVR", "END_CNVR", "MAIN_MODEL", "ALL_MODELS")))


# Residual analysis results; File generated by "07_CNVR/combined/02_residual_analysis.R"
res <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/combined/data/residual_analysis/all_residual_analysis_response.txt", select = c(1:11,13:14)))


# CoxPH analysis results; File generated by "07_CNVR/combined/03_CoxPH.R"
cox <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/combined/data/coxPH/coxPH_top_fit_results.txt"))


# Posthoc Fisher analysis results; File generated by "07_CNVR/combined/04_Fisher.R"
fisher <- as.data.frame(fread("/cauwerx/gwas/07_CNVR/combined/data/Fisher/posthoc_Fisher_cnvrs.txt", header = T))



#################################################
### STEP 2: Summarize data ######################

# Residual analysis
res$RESIDUAL <- 0
res[which(res$P_alt <= thr), "RESIDUAL"] <- 1
cnvrs <- left_join(cnvrs, res[, c(1,2,14)], by = c("PHENO", "TOP_SNP"))

# CoxPH
cox$CoxPH <- 0
cox[which(cox$P_coxPH <= thr), "CoxPH"] <- 1
cnvrs <- left_join(cnvrs, cox[, c(1,2,13)], by = c("PHENO", "TOP_SNP"))

# Posthoc Fisher
fisher$FISHER <- 0
fisher[which(fisher$P_FISHER <= thr), "FISHER"] <- 1
cnvrs <- left_join(cnvrs, fisher[, c(1,2,13)], by = c("PHENO", "TOP_SNP"))



#################################################
### STEP 3: Statistical tiers ###################

cnvrs$TIER <- rowSums(cnvrs[, c("RESIDUAL", "CoxPH", "FISHER")])


#################################################
### Save results ################################

fwrite(cnvrs, "/cauwerx/gwas/07_CNVR/combined/data/combined_CNVR/all_combined_CNVR_annotated.txt", col.names = T, row.names = F, quote = F, sep = "\t")
